// apps/web/src/features/admin/users/user.schema.ts
export const UserFormSchema = z.object({
  userId: z.string().optional(),
  name: z.string().min(1, 'Name is required'),
  firstName: z.string().optional().nullable(),
  lastName: z.string().optional().nullable(),
  email: z.string().email('Invalid email address'),
  role: z.enum(Users.role.enumValues).optional(),
  membershipTier: z.enum(['basic', 'pro', 'enterprise']).optional().nullable(),
})
export type UserFormValues = z.infer<typeof UserFormSchema>